requirements.txt
opencv-python
deepface
pandas
schedule
python-dotenv
matplotlib
tk
Pillow
